+++ src/test/java/kitchenpos/products/domain/ProductTest.java
@@ -4,19 +4,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.math.BigDecimal;
+
+import kitchenpos.Fixtures;
 import kitchenpos.products.application.FakeProfanityValidator;
 import kitchenpos.products.domain.tobe.ProfanityValidator;
 import kitchenpos.products.domain.tobe.Product;
-import kitchenpos.products.fixture.ProductFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-//    - 상품을 생성할 수 있다.
-//    - 상품의 가격을 변경할 수 있다.
-//    - 상품의 가격이 올바르지 않으면 변경할 수 없다.
-//    - 상품의 가격은 0원 이상이어야 한다.
-//    - 상품의 가격이 변경될 때 메뉴의 가격이 메뉴에 속한 상품 금액의 합보다 크면 메뉴가 숨겨진다.
 @DisplayName("상품 테스트")
 class ProductTest {
 
@@ -44,7 +40,7 @@ class ProductTest {
   @DisplayName("상품의 가격을 변경할 수 있다.")
   @Test
   void changeProductPrice(){
-    Product actual = ProductFixture.createProduct();
+    Product actual = Fixtures.createProduct();
 
     actual.changeProductPrice(20_000L);
